backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic-settings
yt-dlp
python-multipart
openai
pyyaml

